combined reducer state types, removed effect hook, undid change to map container type
