+++ src/components/mapPageComponents/ducks/thunks.ts
