+++ src/components/mapPageComponents/mapPage/index.tsx
@@ -1,17 +1,22 @@
 import React from 'react';
 import MapSidebar from '../mapSidebar';
-import MapContainer from '../mapContainer';
+import MapView from '../mapView';
 import MapLayout from '../mapLayout';
 import { Layout } from 'antd';
+import { BlockGeoData, NeighborhoodGeoData } from '../ducks/types';
 
 const { Content, Sider } = Layout;
 
 type MapPageProps = {
+  readonly blocks: BlockGeoData;
+  readonly neighborhoods: NeighborhoodGeoData;
   readonly sidebarHeader: string;
   readonly sidebarDescription: string;
 };
 
 const MapPage: React.FC<MapPageProps> = ({
+  blocks,
+  neighborhoods,
   sidebarHeader,
   sidebarDescription,
   children,
@@ -20,7 +25,7 @@ const MapPage: React.FC<MapPageProps> = ({
     <>
       <MapLayout>
         <Content>
-          <MapContainer />
+          <MapView blocks={blocks} neighborhoods={neighborhoods} />
         </Content>
         <Sider width="20vw">
           <MapSidebar header={sidebarHeader} description={sidebarDescription}>
