+++ src/components/mapPageComponents/mapView/index.tsx
@@ -1,9 +1,9 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import useWindowDimensions, { WindowTypes } from '../../window-dimensions';
 import { Input, message } from 'antd';
 import { Loader } from '@googlemaps/js-api-loader';
 import styled from 'styled-components';
-import { BlockGeoData, NeighborhoodGeoData } from '../mapContainer/ducks/types';
+import { BlockGeoData, NeighborhoodGeoData } from '../ducks/types';
 
 const StyledSearch = styled(Input.Search)`
   width: 40vw;
@@ -61,225 +61,222 @@ const MapView: React.FC<MapViewProps> = ({ blocks, neighborhoods }) => {
     { key: 'Bay Village', value: 'BV' },
   ];
 
-  useEffect(() => {
-    loader.load().then(() => {
-      map = new google.maps.Map(document.getElementById('map') as HTMLElement, {
-        center: BOSTON,
-        zoom: 12,
-        fullscreenControl: false,
-        mapTypeControl: false,
-        // mapId: mapId,
-        restriction: {
-          latLngBounds: BOSTON_BOUNDS,
-          strictBounds: false,
-        },
+  loader.load().then(() => {
+    map = new google.maps.Map(document.getElementById('map') as HTMLElement, {
+      center: BOSTON,
+      zoom: 12,
+      fullscreenControl: false,
+      mapTypeControl: false,
+      // mapId: mapId,
+      restriction: {
+        latLngBounds: BOSTON_BOUNDS,
+        strictBounds: false,
+      },
+    });
+
+    const input = document.getElementById('pac-input') as HTMLInputElement;
+    const autocomplete = new google.maps.places.Autocomplete(input);
+    autocomplete.addListener('place_changed', () => {
+      const place = autocomplete.getPlace();
+
+      if (!place.geometry) {
+        window.alert("No details available for input: '" + place.name + "'");
+        return;
+      }
+
+      // If the place has a geometry, then present it on a map.
+      if (place.geometry.viewport) {
+        map.fitBounds(place.geometry.viewport);
+      } else {
+        map.setCenter(place.geometry.location);
+        map.setZoom(zoomedInLevel);
+      }
+    });
+
+    // Creates a new layer
+    const privateStreetsLayer = new google.maps.Data({ map });
+
+    // Loads the objects into the layer
+    privateStreetsLayer.loadGeoJson(
+      'https://raw.githubusercontent.com/florisdobber/SFTT-map-test/master/private_streets.json',
+    );
+
+    // Sets the style of the layer to simple red lines
+    function setPrivateStreetsStyle(v: boolean) {
+      privateStreetsLayer.setStyle({
+        strokeColor: 'red',
+        strokeWeight: 2,
+        visible: v,
       });
+    }
 
-      const input = document.getElementById('pac-input') as HTMLInputElement;
-      const autocomplete = new google.maps.places.Autocomplete(input);
-      autocomplete.addListener('place_changed', () => {
-        const place = autocomplete.getPlace();
+    // Initially false while the neighborhoods are shown
+    setPrivateStreetsStyle(false);
 
-        if (!place.geometry) {
-          window.alert("No details available for input: '" + place.name + "'");
-          return;
+    // Creates a new layer
+    const blocksLayer = new google.maps.Data({ map });
+
+    // Loads the objects into the layer
+    blocksLayer.addGeoJson(blocks);
+
+    // Sets the style of the layer to colored blocks with black outline
+    function setBlocksStyle(v: boolean) {
+      blocksLayer.setStyle((feature) => {
+        let color = 'green';
+
+        // Use this for coloring reserved/completed blocks a different color
+        if (feature.getProperty('ID') % 10 === 0) {
+          color = 'yellow';
         }
 
-        // If the place has a geometry, then present it on a map.
-        if (place.geometry.viewport) {
-          map.fitBounds(place.geometry.viewport);
-        } else {
-          map.setCenter(place.geometry.location);
-          map.setZoom(zoomedInLevel);
+        if (feature.getProperty('ID') % 10 === 1) {
+          color = 'red';
         }
-      });
 
-      // Creates a new layer
-      const privateStreetsLayer = new google.maps.Data({ map });
+        // Use this for selecting blocks
+        if (feature.getProperty('PRECINCT') === 'YES') {
+          color = 'black';
+        }
 
-      // Loads the objects into the layer
-      privateStreetsLayer.loadGeoJson(
-        'https://raw.githubusercontent.com/florisdobber/SFTT-map-test/master/private_streets.json',
-      );
-
-      // Sets the style of the layer to simple red lines
-      function setPrivateStreetsStyle(v: boolean) {
-        privateStreetsLayer.setStyle({
-          strokeColor: 'red',
-          strokeWeight: 2,
+        // Set styling here
+        return {
+          fillColor: color,
+          strokeColor: 'black',
+          strokeWeight: 1,
           visible: v,
-        });
-      }
+        };
+      });
+    }
 
-      // Initially false while the neighborhoods are shown
-      setPrivateStreetsStyle(false);
+    // Initially false while the neighborhoods are shown
+    setBlocksStyle(false);
 
-      // Creates a new layer
-      const blocksLayer = new google.maps.Data({ map });
+    // Creates a new layer
+    const neighborhoodsLayer = new google.maps.Data({ map });
 
-      // Loads the objects into the layer
-      blocksLayer.addGeoJson(blocks);
-
-      // Sets the style of the layer to colored blocks with black outline
-      function setBlocksStyle(v: boolean) {
-        blocksLayer.setStyle((feature) => {
-          let color = 'green';
-
-          // Use this for coloring reserved/completed blocks a different color
-          if (feature.getProperty('ID') % 10 === 0) {
-            color = 'yellow';
-          }
-
-          if (feature.getProperty('ID') % 10 === 1) {
-            color = 'red';
-          }
-
-          // Use this for selecting blocks
-          if (feature.getProperty('PRECINCT') === 'YES') {
-            color = 'black';
-          }
-
-          // Set styling here
-          return {
-            fillColor: color,
-            strokeColor: 'black',
-            strokeWeight: 1,
-            visible: v,
-          };
-        });
-      }
-
-      // Initially false while the neighborhoods are shown
-      setBlocksStyle(false);
-
-      // Creates a new layer
-      const neighborhoodsLayer = new google.maps.Data({ map });
-
-      // Function: Returns name in shortHand
-      function shortHand(name: string): string {
-        for (const shName of shortHandNames) {
-          if (shName.key === name) {
-            return shName.value;
-          }
+    // Function: Returns name in shortHand
+    function shortHand(name: string): string {
+      for (const shName of shortHandNames) {
+        if (shName.key === name) {
+          return shName.value;
         }
-        return name;
       }
-      // Loads the objects into the layer
-      neighborhoodsLayer.addGeoJson(neighborhoods);
-      /*
-        () => {
-          // For each feature in neighbourhoodsLayer, add a marker
-          // We need to do it here as the GeoJson needs to load first
-          neighborhoodsLayer.forEach((feature) => {
-            // If you want, check here for some constraints.
-            const marker = new google.maps.Marker({
-              map,
-              draggable: false,
-              label: {
-                color: 'white',
-                fontWeight: 'bold',
-                text: shortHand(feature.getProperty('Name')),
-              },
-              // Removed the icon here, only text on map.
-              icon: {
-                labelOrigin: new google.maps.Point(11, 50),
-                url: '',
-                size: new google.maps.Size(22, 40),
-              },
-              position: {
-                lat: feature.getProperty('Latitude'),
-                lng: feature.getProperty('Longitude'),
-              },
-            });
-            markersArray.push(marker);
-            marker.setMap(map);
+      return name;
+    }
+    // Loads the objects into the layer
+    neighborhoodsLayer.addGeoJson(neighborhoods);
+    /*
+      () => {
+        // For each feature in neighbourhoodsLayer, add a marker
+        // We need to do it here as the GeoJson needs to load first
+        neighborhoodsLayer.forEach((feature) => {
+          // If you want, check here for some constraints.
+          const marker = new google.maps.Marker({
+            map,
+            draggable: false,
+            label: {
+              color: 'white',
+              fontWeight: 'bold',
+              text: shortHand(feature.getProperty('Name')),
+            },
+            // Removed the icon here, only text on map.
+            icon: {
+              labelOrigin: new google.maps.Point(11, 50),
+              url: '',
+              size: new google.maps.Size(22, 40),
+            },
+            position: {
+              lat: feature.getProperty('Latitude'),
+              lng: feature.getProperty('Longitude'),
+            },
+          });
+          markersArray.push(marker);
+          marker.setMap(map);
+        });
+      },
+    );
+    */
+
+    // Sets the style of the layer to green shades based on property values with white outline
+    function setNeighborhoodsStyle(v: boolean) {
+      neighborhoodsLayer.setStyle((feature) => {
+        return {
+          fillColor: 'green',
+          fillOpacity: (feature.getProperty('Neighborhood_ID') / 100) * 2 + 0.1, // TODO: replace this with completion percentage
+          strokeWeight: 1,
+          strokeColor: 'white',
+          visible: v,
+        };
+      });
+    }
+
+    function toggleMarkers(v: boolean) {
+      for (const marker of markersArray) {
+        marker.setVisible(v);
+      }
+    }
+
+    // Initially true while the neighborhoods are shown by themselves
+    setNeighborhoodsStyle(true);
+
+    // Check for clicks on neighborhoods and zoom to when clicked on a neighborhood
+    neighborhoodsLayer.addListener('click', (event) => {
+      map.setZoom(14);
+      map.panTo({
+        lat: event.feature.getProperty('Latitude'),
+        lng: event.feature.getProperty('Longitude'),
+      });
+    });
+
+    // Asks user if they want to show their current location
+    if (navigator.geolocation) {
+      navigator.geolocation.getCurrentPosition(
+        (pos) => {
+          const me = new google.maps.LatLng(
+            pos.coords.latitude,
+            pos.coords.longitude,
+          );
+          // eslint-disable-next-line
+          const userLocation = new google.maps.Marker({
+            position: me,
+            map,
+            clickable: false,
+            icon: {
+              path: google.maps.SymbolPath.CIRCLE,
+              scale: 5,
+              fillColor: 'blue',
+              fillOpacity: 1,
+              strokeWeight: 0,
+            },
           });
         },
+        () => {
+          message.info(
+            'Enable access to your location to display where you are on the map.',
+            5,
+          );
+        },
       );
-      */
+    }
 
-      // Sets the style of the layer to green shades based on property values with white outline
-      function setNeighborhoodsStyle(v: boolean) {
-        neighborhoodsLayer.setStyle((feature) => {
-          return {
-            fillColor: 'green',
-            fillOpacity:
-              (feature.getProperty('Neighborhood_ID') / 100) * 2 + 0.1, // TODO: replace this with completion percentage
-            strokeWeight: 1,
-            strokeColor: 'white',
-            visible: v,
-          };
-        });
-      }
+    // Shows or hides layers based on the zoom level
+    function handleZoomChange() {
+      google.maps.event.addListener(map, 'zoom_changed', () => {
+        const zoomLevel = map.getZoom();
+        let zoomedIn = false;
 
-      function toggleMarkers(v: boolean) {
-        for (const marker of markersArray) {
-          marker.setVisible(v);
+        if (zoomLevel >= 13) {
+          zoomedIn = true;
         }
-      }
-
-      // Initially true while the neighborhoods are shown by themselves
-      setNeighborhoodsStyle(true);
-
-      // Check for clicks on neighborhoods and zoom to when clicked on a neighborhood
-      neighborhoodsLayer.addListener('click', (event) => {
-        map.setZoom(14);
-        map.panTo({
-          lat: event.feature.getProperty('Latitude'),
-          lng: event.feature.getProperty('Longitude'),
-        });
+        setNeighborhoodsStyle(!zoomedIn);
+        toggleMarkers(!zoomedIn);
+        setBlocksStyle(zoomedIn);
+        setPrivateStreetsStyle(zoomedIn);
       });
+    }
 
-      // Asks user if they want to show their current location
-      if (navigator.geolocation) {
-        navigator.geolocation.getCurrentPosition(
-          (pos) => {
-            const me = new google.maps.LatLng(
-              pos.coords.latitude,
-              pos.coords.longitude,
-            );
-            // eslint-disable-next-line
-            const userLocation = new google.maps.Marker({
-              position: me,
-              map,
-              clickable: false,
-              icon: {
-                path: google.maps.SymbolPath.CIRCLE,
-                scale: 5,
-                fillColor: 'blue',
-                fillOpacity: 1,
-                strokeWeight: 0,
-              },
-            });
-          },
-          () => {
-            message.info(
-              'Enable access to your location to display where you are on the map.',
-              5,
-            );
-          },
-        );
-      }
-
-      // Shows or hides layers based on the zoom level
-      function handleZoomChange() {
-        google.maps.event.addListener(map, 'zoom_changed', () => {
-          const zoomLevel = map.getZoom();
-          let zoomedIn = false;
-
-          if (zoomLevel >= 13) {
-            zoomedIn = true;
-          }
-          setNeighborhoodsStyle(!zoomedIn);
-          toggleMarkers(!zoomedIn);
-          setBlocksStyle(zoomedIn);
-          setPrivateStreetsStyle(zoomedIn);
-        });
-      }
-
-      handleZoomChange();
-    });
-  }, []);
+    handleZoomChange();
+  });
 
   return (
     <>
