+++ src/containers/landing/index.tsx
@@ -1,4 +1,5 @@
-import React from 'react';
+import React, { useEffect } from 'react';
+import { connect, useDispatch } from 'react-redux';
 import { Helmet } from 'react-helmet';
 import { Typography } from 'antd';
 import MapPage from '../../components/mapPageComponents/mapPage';
@@ -8,10 +9,25 @@ import LandingTreeStats from '../../components/landingTreeStats';
 import useWindowDimensions, {
   WindowTypes,
 } from '../../components/window-dimensions';
+import { AsyncRequestKinds } from '../../utils/asyncRequest';
+import { getMapGeoData } from '../../components/mapPageComponents/ducks/thunks';
+import { MapGeoDataReducerState } from '../../components/mapPageComponents/ducks/types';
+import { C4CState } from '../../store';
 
 const { Paragraph } = Typography;
 
-const Landing: React.FC = () => {
+interface LandingProps {
+  readonly neighborhoods: MapGeoDataReducerState['neighborhoodGeoData'];
+  readonly blocks: MapGeoDataReducerState['blockGeoData'];
+}
+
+const Landing: React.FC<LandingProps> = ({ neighborhoods, blocks }) => {
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch(getMapGeoData());
+  }, [dispatch]);
+
   const { windowType } = useWindowDimensions();
 
   const statMoneySaved = 100000;
@@ -29,15 +45,21 @@ const Landing: React.FC = () => {
               content="The first page someone sees if they are not logged in, contains a read only map of Boston neighborhoods and some information about the tree counts of Speak for the Trees"
             />
           </Helmet>
-          <MobileMapPage>
-            <MobileLandingBar>
-              <LandingTreeStats
-                moneySaved={statMoneySaved}
-                rainWater={statRainWater}
-                carbonEmissions={statCarbonEmissions}
-              />
-            </MobileLandingBar>
-          </MobileMapPage>
+          {blocks.kind === AsyncRequestKinds.Completed &&
+            neighborhoods.kind === AsyncRequestKinds.Completed && (
+              <MobileMapPage
+                neighborhoods={neighborhoods.result}
+                blocks={blocks.result}
+              >
+                <MobileLandingBar>
+                  <LandingTreeStats
+                    moneySaved={statMoneySaved}
+                    rainWater={statRainWater}
+                    carbonEmissions={statCarbonEmissions}
+                  />
+                </MobileLandingBar>
+              </MobileMapPage>
+            )}
         </>
       );
 
@@ -53,16 +75,21 @@ const Landing: React.FC = () => {
               content="The first page someone sees if they are not logged in, contains a read only map of Boston neighborhoods and some information about the tree counts of Speak for the Trees"
             />
           </Helmet>
-          <MapPage
-            sidebarHeader="Boston's Street Trees"
-            sidebarDescription="Dreamcatcher kogi taiyaki keytar. Swag typewriter craft beer cronut pok pok gentrify flannel salvia deep v pork belly pitchfork. Swag fashion axe fam. Occupy biodiesel jean shorts affogato PBR&B freegan bushwick vegan four loko pickled."
-          >
-            <LandingTreeStats
-              moneySaved={statMoneySaved}
-              rainWater={statRainWater}
-              carbonEmissions={statCarbonEmissions}
-            />
-          </MapPage>
+          {blocks.kind === AsyncRequestKinds.Completed &&
+            neighborhoods.kind === AsyncRequestKinds.Completed && (
+              <MapPage
+                neighborhoods={neighborhoods.result}
+                blocks={blocks.result}
+                sidebarHeader="Boston's Street Trees"
+                sidebarDescription="Dreamcatcher kogi taiyaki keytar. Swag typewriter craft beer cronut pok pok gentrify flannel salvia deep v pork belly pitchfork. Swag fashion axe fam. Occupy biodiesel jean shorts affogato PBR&B freegan bushwick vegan four loko pickled."
+              >
+                <LandingTreeStats
+                  moneySaved={statMoneySaved}
+                  rainWater={statRainWater}
+                  carbonEmissions={statCarbonEmissions}
+                />
+              </MapPage>
+            )}
         </>
       );
 
@@ -71,4 +98,11 @@ const Landing: React.FC = () => {
   }
 };
 
-export default Landing;
+const mapStateToProps = (state: C4CState): LandingProps => {
+  return {
+    neighborhoods: state.mapGeoDataState.neighborhoodGeoData,
+    blocks: state.mapGeoDataState.blockGeoData,
+  };
+};
+
+export default connect(mapStateToProps)(Landing);
